CHROMIUM: atmel: Add support for probing device from ACPI
With firmware support it is possible to discover and probe devices
on I2C buses without needing the glue layer in chromeos_laptop that
we use today.

This requires that the drivers implement the new acpi_match_table
with compatible ACPI IDs.

I chose to use ATML0000 for touchpad and ATML0001 for touchscreen
since the driver likes to determine which has been probed based on
the device id and our config/firmware update scripts need to easily
distinguish between touchpad and touchscreen.

When probed from ACPI the i2c device id structure is NULL so we have
to pull out the ACPI ID and compare the ID to determine if it is a
touchpad or touchscreen.

Additionally the driver check to see if something is found at the
specified address is changed to always run, rather than just when
probing from DeviceTree.

BUG=chrome-os-partner:24380
TEST=manual:

With firmware changes to expose I2C controllers as ACPI devices
instead of in PCI mode and expose the Touchpad as an ACPI device
in that controller namespace with an appropriate I2cSerialBus()
resource the existing touchpad is probed without using the glue
layer in chromeos_laptop:

input: Atmel maXTouch Touchpad as /devices/platform/80860F41:00/i2c-0/0-004b/input/input3

The interrupt is also probed properly:
  51:        321          0   IO-APIC-edge      ATML0000:01

Change-Id: I8247f150fcd62037d8df6d39993760504828a261
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/182330
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>

[benzh: 3.14 rebase. Resolved trivial conflicts]
Signed-off-by: Ben Zhang <benzh@chromium.org>