CHROMIUM: ARM: tegra: dvfs: update set rate and add get_rate
Update set_rate to just return 0 if a clock is not subject to DVFS.
Also remove the warning. Implement a new get_rate() function which
returns the current clock rate as used for determining the required
voltage. This is used by is_clk_prepared() to determine if a clock is
used or not. We can't use clk_enabled() for this because
clk_disable_unused will disable unused clocks before unpreparing them,
so those clocks would stay prepared and contribute to the minimum
voltage requirement.

BUG=None
TEST=None

Change-Id: I98dd50d8cc53f820a8dadd7e9a14b118e60baf96
Signed-off-by: Peter De Schrijver <pdeschrijver@nvidia.com>
Signed-off-by: Bill Huang <bilhuang@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/179307
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>