UPSTREAM: drm/i915: Defer request freeing
With context destruction, we always want to be able to tear down the
underlying address space. This is invoked on the last unreference to the
context which could happen before we've moved all objects to the
inactive list. To enable a clean tear down the address space, make sure
to process the request free lastly.

Without this change, we cannot guarantee to we don't still have active
objects in the VM.

As an example of a failing case:
CTX-A is created, count=1
CTX-A is used during execbuf
	does a context switch count = 2
	and add_request count = 3
CTX B runs, switches, CTX-A count = 2
CTX-A is destroyed, count = 1
retire requests is called
	free_request from CTX-A, count = 0 <--- free context with active object

As mentioned above, by doing the free request after processing the
active list, we can avoid this case.

Signed-off-by: Ben Widawsky <ben@bwidawsk.net>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>

BUG=None
TEST=With entire patch series applied, RPS enables higher framerate on bwtm2

Change-Id: Icdce2e30f9da2f6946f8d2e69bf809531ff06845
Link: git://anongit.freedesktop.org/drm-intel br:drm-intel-next
(cherry picked from commit e20780439b26ba95aeb29d3e27cd8cc32bc82a4c)
Reviewed-on: https://chromium-review.googlesource.com/191564
Reviewed-by: James Ausmus <james.ausmus@intel.com>
Tested-by: James Ausmus <james.ausmus@intel.com>
Reviewed-by: Joe Konno <joe.konno@intel.com>
Tested-by: Joe Konno <joe.konno@intel.com>
Reviewed-by: Ben Zhang <benzh@chromium.org>
Commit-Queue: Joe Konno <joe.konno@intel.com>