CHROMIUM: cpufreq: tegra124: add cluster-switching support
Add cluster-switching support to the Tegra124 cpufreq driver.

Switching to the LP cluster will occur when only one CPU is online
and the requested frequency is within the range of the LP cluster.

Switching back to the G cluster will occur when a secondary CPU is
preparing to come up or the requested frequency is outside of the
LP cluster's limits.

BUG=None
TEST=Boot on Nyan rev1 and see cluster switching happen as CPUs
are loaded and unloaded.

Change-Id: Ia3dfd117698d19e17a63873b15833dde75f18e63
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/179230
Reviewed-by: Dylan Reid <dgreid@chromium.org>