PCI: Export MSI mode in sysfs using attributes, not kobjects
The PCI MSI sysfs code is a mess with kobjects for things that don't
really need to be kobjects.  This patch creates attributes dynamically
for the MSI interrupts instead of using kobjects.

[bhelgaas: This also fixes a problem reported by Veaceslav (see link
below): calling pci_enable_msi() too soon after pci_disable_msi() can fail
with "sysfs: cannot create duplicate filename '.../msi_irqs' because the
msi_irqs kset was still active.  This can happen with various module
unload/reload and bonding scenarios.]

Note, this removes a directory from the current MSI interrupt sysfs
code:

  old MSI kobjects:
    pci_device
       └── msi_irqs
           └── 40
               └── mode

  new MSI attributes:
    pci_device
       └── msi_irqs
           └── 40

As there was only one file "mode" with the kobject model, the interrupt
number is now a file that returns the "mode" of the interrupt (msi vs.
msix).

[bhelgaas: Neil confirmed that this sysfs change doesn't affect irqbalance
because it only cares about the *names* of the entries in msi_irqs/; it
doesn't care whether those entries are files or directories.  It may break
other programs that depend on sysfs layout.]

Link: http://lkml.kernel.org/r/1379382464-7920-1-git-send-email-vfalico@redhat.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Bjorn Helgaas <bhelgaas@google.com>
Reviewed-by: Veaceslav Falico <vfalico@redhat.com>
Acked-by: Neil Horman <nhorman@tuxdriver.com>