CHROMIUM: mmc: tegra: add DT binding to disable runtime PM on specific host
Currently the Tegra SDHC driver supports runtime PM function on each host
by default and can't be disabled. We add a binding in DT to specify if
the host wants to turn it off. This may happen when the SD device that
was attached to the host didn't support runtime PM.

BUG=chrome-os-partner:24498
TEST=boot the device

Change-Id: I45abf551bed0c76c481a69c94d9257d29d471015
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/179810
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>