Enable atmel 540s trackpad for atmel_mxt_ts driver
The Atmel 540s chip has an additional object in Atmel's object protocol
that was not used by previous ChromeOS devices.  This object T100 needs
to be supported if we are to use 540s touchpads in the future.  This CL
adds support to the atmel_mxt_ts drive for the T100 object and its
related parameters so that 540s touchpads will work.

BUG=chrome-os-partner:24559
TEST=Run evtest on new trackpad. Check the pointer moving and that
input events are arriving as expected.  This patched driver was also
tested on a Spring to make sure that the old Atmel pads that don't use
the new T100 format still work.

Change-Id: I4c4512b2b3c1f334196ad672cb17f7db16d293ee
Signed-off-by: Jongpil Jung <jongpil19.jung@samsung.com>
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/180151
Reviewed-by: Wonjoon Lee <woojoo.lee@samsung.com>
Tested-by: Wonjoon Lee <woojoo.lee@samsung.com>
Signed-off-by: Yufeng Shen <miletus@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/188876
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/189536