Don't select EFI from certain special ACPI drivers
Commit 7ea6c6c15ee1 ("Move cper.c from drivers/acpi/apei to
drivers/firmware/efi") results in CONFIG_EFI being enabled even when
the user doesn't want this. Since ACPI APEI used to build fine without
UEFI (and as far as I know also has no functional depency on it), at
least in that case using a reverse dependency is wrong (and a straight
one isn't needed).

Whether the same is true for ACPI_EXTLOG I don't know - if there is a
functional dependency, it should depend on EFI rather than selecting
it. It certainly has (currently) no build dependency.

Adjust Kconfig and build logic so that the bad dependency gets avoided.

Signed-off-by: Jan Beulich <jbeulich@suse.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>