CHROMIUM: video: tegra: host: Add waits to FECS reset
Based on discussion with hw folks, the correct way to wait reset
is to first read back the reset register and then wait a moment. This
should be done also after we release the reset to ensure that the hw
is ready before we actually do anything with it.

BUG=None
TEST=System boots OK

Change-Id: I70d2af0ae745f6a8579bcb89c8165fe6bec9dd22
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/192026
Tested-by: Mark Zhang <markz@nvidia.com>
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>