UPSTREAM: thermal: samsung: change base_common to more meaningful base_second
On Exynos5440 and Exynos5420 there are registers common
across the TMU channels.

To support that, we introduced a ADDRESS_MULTIPLE flag in the
driver and the 2nd set of register base and size are provided
in the "reg" property of the node.

As per Amit's suggestion, this patch changes the base_common
to base_second and SHARED_MEMORY to ADDRESS_MULTIPLE.

BUG=None
TEST=Snow works better

Change-Id: I9c2c1874e64cd20a3c391501d080cd09c2566aaa
Signed-off-by: Naveen Krishna Chatradhi <ch.naveen@samsung.com>
Acked-by: Amit Daniel Kachhap <amit.daniel@samsung.com>
Reviewed-by: Bartlomiej Zolnierkiewicz <b.zolnierkie@samsung.com>
Reviewed-by: Tomasz Figa <t.figa@samsung.com>
Signed-off-by: Eduardo Valentin <edubezval@gmail.com>
Signed-off-by: Doug Anderson <dianders@chromium.org>
(cherry picked from commit 9025d563cd9bd141a7b7f2095b6a760cd9d83a4e)
Reviewed-on: https://chromium-review.googlesource.com/212501
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>