FogBugz #173185: OCRAM ECC addition for Altera SOCFPGA.
Add On-Chip RAM ECC to the Error Detection And Correction module for
tracking ECC errors in Altera's SOCFPGA OCRAM.
This patch adds code to the kernel startup that initializes OCRAM
and ensures the ECC is enabled. The generic EDAC Manager driver
handles the addition of OCRAM to the driver when the OCRAM "compatible"
field in the device tree matches.
Specific data is loaded depending upon which type of ECC (L2 or OCRAM)
is being initalized.

V2 Changes:
- Fix error message in OCRAM initialization.
- Document the device tree bindings for OCRAM & OCRAM ECC.
- Change variable name from sram to iram for clarity.
- Resync with main since EDAC L2 trigger was added.

V3 Changes:
- Remove OCRAM device tree binding file.
- Add conditional compile statements for L2 clear in socfpga.c

V4 Changes:
- Add OCRAM only config.
- Split OCRAM initialization into own files (ocram.c & ocram.h)
- Conditionally compile OCRAM matching tag.

V5 Changes:
- Cleanup of ocramedac address in socfpga.dtsi
- Use pr_err() consistently in ocram.c

V6 Changes:
- Cleanup of ocramedac address in socfpga-ocram-ecc.txt.
- Remove unused .setup function pointer.

Signed-off-by: Thor Thayer <tthayer@altera.com>

Conflicts:

	arch/arm/mach-socfpga/socfpga.c