diag: Increase the maximum number of SSIDs per msg mask range
Diag was using a max ssid range of 100. This has increased due
to addition of new SSIDs. This patch increases the range to 200
and adds checks to limit the SSID masks to this range. This
changes the order of sending the feature masks, sending it before
the msg masks to let the peripherals know the features Apps
support.

Change-Id: Id124a29b38f9a0640cee77cfd49bfbb88c18dcb0
CRs-fixed: 576183
Signed-off-by: Sridhar Ancha <sancha@codeaurora.org>