x86, irq: Change check_vectors() to check_irq_vectors_for_cpu_disable()
In the context of smpboot.c especially, it is not at all obvious what
"check_vectors" actually mean.  Since this function is only called in
one place, without arguments, we can afford to be a bit more verbose
with the function name and make it clear at a glance what the function
does.

Cc: Prarit Bhargava <prarit@redhat.com>
Link: http://lkml.kernel.org/r/1387554609-9823-1-git-send-email-prarit@redhat.com
Signed-off-by: H. Peter Anvin <hpa@linux.intel.com>