drm/exynos: exclude modes from adjustment comparison
The mode adjustment logic is checking whether a "native" mode with the same
resolution as adjusted mode already exists. But it can be an another
adjusted mode (possibly with different refresh rate) or a mode with
unsupported pixel clock. This change introduces a private flag
EXYNOS_MODE_ADJUSTED which is set when the mode is adjusted and skips the
comparison with adjusted modes and also checks whether the comparison target
mode is valid using check_mode(). This solves problem with two scenarios.
One, already encountered is when 1024x768@75 is adjusted to 1024x720@75 and
prevents 1024x768@60 from also being adjusted to 1024x720@60 with both of
them being pruned later. Second is where "native" mode with unsupported
pixel clock could prevent of adjustment of another mode to the same
resolution.

BUG=chromium:331423
TEST=connect Daisy to HP LP2465 monitor. 1024x720 mode should be available.

Change-Id: I8e3fcab58ebe991f556cefe56e228399cbe793e5
Signed-off-by: Dominik Behr <dbehr@chromium.org>
Previous-Reviewed-on: https://chromium-review.googlesource.com/181420
(cherry picked from commit 00618d731cadd818d39fdde4c0a9e159467774b6)
Reviewed-on: https://chromium-review.googlesource.com/182218
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>