FogBugz #175951: nios2: Move the MMU initialization
The current MMU initialization consists only in the flush of all the TLB
entries. This is done based in the cpuinfo values that specify the TLB
geometry. Hence, the current calling of mmu_init() *before* the cpuinfo is
setup is wrong and doesn't perform any TLB flushing.

Fix it by moving the call to mmu_init() to be done *after* the cpuinfo has been
correctly filled.

This commit fixes soft-reboots, which are currently broken. Upon soft-reboot
(from a system that has entered userspace) the TLB would be filled and
were previously never flushed. Such systems booting from a soft-reboot
would fail to enter userpace, panic'ing like this:

  Kernel panic - not syncing: Attempted to kill init! exitcode=0x0000000b

Signed-off-by: Ezequiel Garcia <ezequiel@vanguardiasur.com.ar>
Signed-off-by: Ley Foon Tan <lftan@altera.com>