CHROMIUM: config: Change config files to match wireless-3.8
This change removes some of the configs that don't exist in wireless-3.8 and
turns on some configs that were left out. ATH9K_RATE_CONTROL is set in the
config files for the 3.8 kernel, so this adds it back. This also turns the
EXPERIMENTAL config back on since that is needed for some of the wireless-3.8
configs.

BUG=None
TEST=emerge-rambi chromeos-kernel-3_10

Change-Id: I0afd2b6c9068e423729d28fc6b669cfdcf1fa424
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/181685
Reviewed-by: Olof Johansson <olofj@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>