CHROMIUM: thermal: tegra: fix the num_of_cap_clocks
Since we added gpu clk in the cap clocks array, the num_of_cap_clocks
must consider it. In current codes, the num_of_cap_clocks is wrong,
and will get the wrong clk freq.  This change fixes this issue.

BUG=None
TEST=Verified. After applying the pid governor patch, and heat the
system, the throttle function will be trigged, check the cur_state
and cap clocks in sysfs, the values are expected.

Change-Id: I6aed621de6f33330b9a97d0b41125a6e4facdd37
Signed-off-by: Wei Ni <wni@nvidia.com>
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/181725