USB: gadget: IPA disconnect on USB suspend when remote wakeup is disabled
When remote wakeup feature is disabled due to the usb host decision,
our tethering device is not allowed to wakeup the usb bus after bus
suspend. Therefore, in case the IPA is willing to wakeup the USB bus,
the USB cannot wakeup the host and therefore cannot grant the consumer
resource back to the IPA resource manager. In this case the IPA resource
manager is waiting until expiring timeout which lead system crash.

Therefore, in case of remote wakeup disabled, we should avoid the
expectation from the IPA side to get a response from the USB. So, the
solution is to indicate to the IPA driver that USB cable has been unplugged
when bus suspend with remote wakeup disabled.

CRs-Fixed: 571703
Change-Id: I03e1e2ad3549b7aad0df263de41b9a3dcc2f3503
Signed-off-by: Danny Segal <dsegal@codeaurora.org>