CHROMIUM: Add a new iTCO vendor support option for ChromeOS to disable TCO SMI
This new vendor support is selected by a kernel command line parameter on
x86/x86_64 Chrome OS. It disables TCO SMI when the watchdog timer is enabled.
On some existing devices, this workaround is needed because the TCO SMI handler
in coreboot clears the TIMEOUT bit of TCO1_STS register, preventing the machine
from rebooting when the timer expires. This workaround makes the watchdog work
on existing devices, but it is not necessary for devices with updated firmware.

BUG=chromium:321832
TEST=verified that the Intel TCO watchdog works on peppy

Change-Id: I34fcd169e95a1cad225d8f6233eddf188e9bf177
Signed-off-by: Ben Zhang <benzh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/182058
Reviewed-by: Grant Grundler <grundler@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>