UPSTREAM: usb: cdc-wdm: resp_count can be 0 even if WDM_READ is set
Do not decrement resp_count if it's already 0.

We set resp_count to 0 when the device is closed.  The next open and
read will try to clear the WDM_READ flag if there was leftover data
in the read buffer. This fix is necessary to prevent resubmitting
the read URB in a tight loop because resp_count becomes negative.

The bug can easily be triggered from userspace by not reading all
data in the read buffer, and then closing and reopening the chardev.

Fixes: 8dd5cd5395b9 ("usb: cdc-wdm: avoid hanging on zero length reads")
Cc: <stable@vger.kernel.org> # 3.13
Signed-off-by: Bjørn Mork <bjorn@mork.no>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit f563926fed982f26b391ca42493f55f2447f1b0a)

BUG=chromium:326559
TEST=Tested the following:
1. `USE='kernel-3_10 -kernel-next' emerge-peppy chromeos-kernel-3_10`
2. Manually test cellular support on a DUT with a MBIM modem.
3. Run network_3GIdentifiers, network_3GModemPresent and
   network_3GSmokeTest.

Change-Id: I394d695330ea99ebbeb24b1f95a4744643c6dc3e
Signed-off-by: Ben Chan <benchan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/186758
Reviewed-by: Paul Stewart <pstew@chromium.org>