Revert "CHROMIUM: clk: tegra: remove unnecessary is_prepared() callbacks"
It turns out these callbacks *are* necessary for tegra_dvfs to correctly
ignore all unprepared and disabled clocks.  Suppose a clock is enabled
in hardware at boot, but never prepared by the kernel.  tegra_config_dvfs
will then use the clock's rate to determine core voltage, but the kernel
will later disable it in clk_disable_unused since it has no consumers.
Since the kernel never touches the clock again, we'll never disable
DVFS on it and it's initial rate will always contribute to core voltage.
This may result in the core voltage being held higher than what is
necessary.  To resolve this, we want clk_disable_unused to also
unprepare (and thus disable DVFS on) these unused clocks.  Add back
the is_prepared callbacks so that a clock is considered prepared if
it has either been explicitly prepared or it has DVFS enabled on it.

This reverts commit 6cbe423b607dc1b676e8eee0d404b72f01277251.

BUG=None
TEST=VDD_CORE drops from 1.1V to 0.8V at idle.

Change-Id: Ia56d71e0aa2a7aa5d9b0604bb6f30566e4df5005
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/182561
Reviewed-by: Bill Huang <bilhuang@nvidia.com>
Reviewed-by: Dylan Reid <dgreid@chromium.org>