CHROMIUM: ARM: tegra: DVFS resume fixes
Do not consider clocks which are not enabled and do not have DVFS
enabled on them (which, at this point, is equivalent to the clock
not being prepared) when re-calculating core voltage during resume.
Furthermore, do not update core voltage until all clocks have been
accounted for.  This prevents us from temporarily under-volting a
clock before we've recorded its frequency.

BUG=None
TEST=Run LP0 suspend/resume; check that DVFS is only enabled on the
clocks it was enabled on before suspend.

Change-Id: I42ebbe974e8b826f398cbac8e15c8bc22f4414c0
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/182563
Reviewed-by: Dylan Reid <dgreid@chromium.org>