UPSTREAM: MIPS: Introduce _EXT assembler macro
This patch adds a simple macro to wrap the ext instruction which was
introduced with MIPSR2, and fall back to a shift & and pair for
pre-MIPSR2 CPUs. This will be used in a subsequent patch.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>
Cc: linux-mips@linux-mips.org
Patchwork: https://patchwork.linux-mips.org/patch/6358/
Signed-off-by: Ralf Baechle <ralf@linux-mips.org>
(cherry picked from commit 1d68808748afe21ac35bedc2726043bea33a1c0d)

BUG=None
TEST=Build and boot on Malta

Change-Id: Iba1bf80e8eefe6f57719c003c02aad520dbf31b6
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/207912
Reviewed-by: Olof Johansson <olofj@chromium.org>