[media] adv7180: Prepare for multi-chip support
The adv7180 is part of a larger family of device, which have all a very similar
register map layout. This patch prepares the adv7180 driver for support for
multiple different devices. For now the only difference we care about is the
number of input channel configurations. Also the way the input format is
configured slightly differs between some devices.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>